🔧 move test files and add jest configurations
- test 파일 (*.spec.ts)을 server/test 폴더로 이동하고
  package.json jest 설정의 rootDir과 testRegex를 수정
- package.json jest 설정에 verbose, collectCoverage 추가
